Migrate serialization logic to connector serializer
